requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
cryptography